requirements-dev.txt
cassandra-driver
jupyter
pandas
scikit-learn
tensorflow
boto3
pypyr
python-dotenv
fastapi
uvicorn
gunicorn